refactor: move simulator utils functions
