+++ tests/cli/test_twin_tx.py
@@ -6,12 +6,12 @@ from structlog.testing import capture_logs
 
 from hathor.cli.twin_tx import create_parser, execute
 from hathor.conf import HathorSettings
+from hathor.simulator.utils import add_new_blocks
 from hathor.transaction import Transaction, TransactionMetadata
 from hathor.util import json_loadb
 from tests import unittest
 from tests.utils import (
     add_blocks_unlock_reward,
-    add_new_blocks,
     add_new_transactions,
     execute_mining,
     execute_tx_gen,
