+++ tests/consensus/test_soft_voided2.py
@@ -1,9 +1,10 @@
 from hathor.conf import HathorSettings
 from hathor.graphviz import GraphvizVisualizer
 from hathor.simulator import Simulator
+from hathor.simulator.utils import gen_new_tx
 from tests import unittest
 from tests.simulation.base import SimulatorTestCase
-from tests.utils import BURN_ADDRESS, add_custom_tx, gen_new_tx
+from tests.utils import BURN_ADDRESS, add_custom_tx
 
 settings = HathorSettings()
