+++ tests/p2p/test_split_brain.py
@@ -4,9 +4,10 @@ from mnemonic import Mnemonic
 from hathor.daa import TestMode
 from hathor.graphviz import GraphvizVisualizer
 from hathor.simulator import FakeConnection
+from hathor.simulator.utils import add_new_block
 from hathor.wallet import HDWallet
 from tests import unittest
-from tests.utils import add_blocks_unlock_reward, add_new_block, add_new_double_spending, add_new_transactions
+from tests.utils import add_blocks_unlock_reward, add_new_double_spending, add_new_transactions
 
 
 class BaseHathorSyncMethodsTestCase(unittest.TestCase):
