+++ tests/resources/transaction/test_transaction_confirmation.py
@@ -1,9 +1,10 @@
 from twisted.internet.defer import inlineCallbacks
 
+from hathor.simulator.utils import add_new_blocks
 from hathor.transaction.resources import TransactionAccWeightResource
 from tests import unittest
 from tests.resources.base_resource import StubSite, _BaseResourceTest
-from tests.utils import add_blocks_unlock_reward, add_new_blocks, add_new_transactions
+from tests.utils import add_blocks_unlock_reward, add_new_transactions
 
 
 class BaseTransactionTest(_BaseResourceTest._ResourceTest):
