+++ tests/tx/test_indexes4.py
@@ -1,9 +1,10 @@
 from hathor.crypto.util import decode_address
+from hathor.simulator.utils import add_new_blocks, gen_new_tx
 from hathor.transaction import Transaction
 from hathor.transaction.storage import TransactionMemoryStorage
 from hathor.wallet.base_wallet import WalletOutputInfo
 from tests import unittest
-from tests.utils import add_blocks_unlock_reward, add_new_blocks, gen_new_tx
+from tests.utils import add_blocks_unlock_reward
 
 
 class BaseSimulatorIndexesTestCase(unittest.TestCase):
