+++ tests/tx/test_mining.py
@@ -2,10 +2,10 @@ from typing import Any
 
 from hathor.conf import HathorSettings
 from hathor.mining import BlockTemplate
+from hathor.simulator.utils import add_new_blocks
 from hathor.transaction import Block, sum_weights
 from hathor.transaction.storage import TransactionMemoryStorage
 from tests import unittest
-from tests.utils import add_new_blocks
 
 settings = HathorSettings()
