+++ tests/tx/test_tips.py
@@ -1,12 +1,7 @@
+from hathor.simulator.utils import add_new_block, add_new_blocks
 from hathor.transaction import Transaction
 from tests import unittest
-from tests.utils import (
-    add_blocks_unlock_reward,
-    add_new_block,
-    add_new_blocks,
-    add_new_double_spending,
-    add_new_transactions,
-)
+from tests.utils import add_blocks_unlock_reward, add_new_double_spending, add_new_transactions
 
 
 class BaseTipsTestCase(unittest.TestCase):
