+++ tests/wallet/test_balance_update.py
@@ -1,11 +1,12 @@
 from hathor.conf import HathorSettings
 from hathor.crypto.util import decode_address
+from hathor.simulator.utils import add_new_blocks
 from hathor.transaction import Transaction, TxInput, TxOutput
 from hathor.transaction.scripts import P2PKH
 from hathor.wallet.base_wallet import SpentTx, UnspentTx, WalletBalance, WalletInputInfo, WalletOutputInfo
 from hathor.wallet.exceptions import PrivateKeyNotFound
 from tests import unittest
-from tests.utils import add_blocks_unlock_reward, add_new_blocks, create_tokens
+from tests.utils import add_blocks_unlock_reward, create_tokens
 
 settings = HathorSettings()
