+++ tests/wallet/test_wallet.py
@@ -6,13 +6,14 @@ from cryptography.hazmat.primitives import serialization
 
 from hathor.conf import HathorSettings
 from hathor.crypto.util import decode_address, get_address_b58_from_public_key, get_private_key_bytes
+from hathor.simulator.utils import add_new_block
 from hathor.transaction import Transaction, TxInput
 from hathor.wallet import Wallet
 from hathor.wallet.base_wallet import WalletBalance, WalletInputInfo, WalletOutputInfo
 from hathor.wallet.exceptions import InsufficientFunds, InvalidAddress, OutOfUnusedAddresses, WalletLocked
 from hathor.wallet.keypair import KeyPair
 from tests import unittest
-from tests.utils import add_blocks_unlock_reward, add_new_block, create_tokens, get_genesis_key
+from tests.utils import add_blocks_unlock_reward, create_tokens, get_genesis_key
 
 settings = HathorSettings()
